Fix func builtin in kretprobes
For ret probes, the IP register will not be pointing to the function we
want to trace. It may point to a kernel trampoline, or it may point to
the caller of the traced function, as it fires after the "ret"
instruction has executed.

Switch to using the get_func_ip BPF helper in kret probes.

Disable func in kretprobes when this helper is not available, and for
uretprobes everywhere, as it previously returned incorrect results.