CMake: Have "embed" rule depend on source files directly instead of provided target argument
The dependency chain is as follows (using data_source/BTF as an example
but applies to all targets and formats):
1. data_source.c
2. data_source.o
3. data_source.btf
4. data_source_btf.hex
5. data_source_btf.h
6. bpftrace_test

An update to (1) triggers rebuilds of (2) and (3), but not (4) or later.

This change makes it so that (4) and also later targets are rebuilt
when the source files change.

Tested with Make and Ninja.

Signed-off-by: Alastair Robertson <ajor@meta.com>